requirements.txt
# requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytz>=2023.0
python-dateutil>=2.8.0
newsapi-python
alpha_vantage
nsepython
nsetools
deepseek
yfinance>=0.2.18
kiteconnect>=4.0.0
# pandas-ta not needed - implementing indicators manually with pandas/numpy
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
websocket-client>=1.6.0
